// client/src/generated/client/models/RsUser.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Play with friends
 * Play with friends
 *
 * The version of the OpenAPI document: 1.0
 * Contact: theo.dm94@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RsUser
 */
export interface RsUser {
    /**
     * 
     * @type {string}
     * @memberof RsUser
     */
    userName: string;
    /**
     * 
     * @type {string}
     * @memberof RsUser
     */
    userPublicID: string;
    /**
     * 
     * @type {string}
     * @memberof RsUser
     */
    userKey: string;
}

export function RsUserFromJSON(json: any): RsUser {
    return RsUserFromJSONTyped(json, false);
}

export function RsUserFromJSONTyped(json: any, ignoreDiscriminator: boolean): RsUser {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'userName': json['userName'],
        'userPublicID': json['userPublicID'],
        'userKey': json['userKey'],
    };
}

export function RsUserToJSON(value?: RsUser | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'userName': value.userName,
        'userPublicID': value.userPublicID,
        'userKey': value.userKey,
    };
}